feat: correctly call functions with default argument values
Some functions have default argument values defined. To handle that
correctly we assume to pass (by order) the rest of the arguments the
user did not pass.

D doesn't have (yet) the support for named arguments. This means we can
assume the values are always passed by the same order they are defined
in their function signature. Assuming the user has following function:

void fun(int a = 4, int b) {}

The variable 'b' is mandatory, and to write to it, the user has to also
write to 'a' even though it is an optional argument. Knowing this, we
are safe to assume that the indices of the default values and the
quantity of actual arguments passed by the user comply with each other.

Note: void is used for any type that does not have a default value
resulting in an error, not the best error message but it's better than
nothing;

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>