fix: correctly check for a valid identifier
Currently the function does not allow unicode characters which is not
conformant with the D spec.

Moreover the verification is weak which can lead security issues. The
user can pass multiple tokens and the verification might be valid:

```d
static assert(isDecoratedName!"_; alias myNewAlias");
```

This is also fixed with the new algorithm.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>